Add files via upload
This lab was designed to simulate the execution of ARM instructions in a C environment using the <stdio.h> library.  The instructions given are to have a file be input into the program and be read into a memory array, at the address of 200.  The program needs to use arrays to parse the string input, convert the input into an address, or a usable value of some sort, and follow the arithmetic or logical instructions given by the input, to then store and(/)or load addresses to memory locations all with the implementation of a program counter.